# internal/openapi/meetup-types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.1 DO NOT EDIT.
package openapi

import (
	"time"
)

const (
	BasicAuthScopes = "BasicAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	// Error code
	Code int32 `json:"Code"`

	// Error message
	Message string `json:"Message"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error Error `json:"Error"`
}

// Meetup defines model for Meetup.
type Meetup struct {
	AttendeeList []string `json:"AttendeeList"`

	// username of organiser.
	Organiser string `json:"Organiser"`

	// Meetup Date and Time
	PlannedDateTime time.Time `json:"PlannedDateTime"`

	// Meetup Reminder Date and Time
	RemindDateTime *time.Time `json:"RemindDateTime,omitempty"`
}

// MeetupResponse defines model for MeetupResponse.
type MeetupResponse struct {
	Meetup Meetup `json:"Meetup"`

	// Unique ID of the created meetup
	MeetupID string `json:"MeetupID"`
}

// CreateMeetupJSONBody defines parameters for CreateMeetup.
type CreateMeetupJSONBody Meetup

// CreateMeetupJSONRequestBody defines body for CreateMeetup for application/json ContentType.
type CreateMeetupJSONRequestBody CreateMeetupJSONBody